+++ src/cowl_iterator.c
@@ -25,23 +25,23 @@ static bool cowl_count(void *count, cowl_unused void *obj) {
     return true;
 }
 
-CowlIterator cowl_iterator_vec_init(UVec(CowlObjectPtr) *vec) {
+CowlIterator cowl_iterator_vec(UVec(CowlObjectPtr) *vec) {
     if (!vec) {
         vec = ulib_alloc(vec);
-        if (vec) *vec = uvec_init(CowlObjectPtr);
+        if (vec) *vec = uvec(CowlObjectPtr);
     }
     return (CowlIterator) { .ctx = vec, .for_each = cowl_store_vec };
 }
 
-CowlIterator cowl_iterator_set_init(UHash(CowlObjectTable) *set) {
+CowlIterator cowl_iterator_set(UHash(CowlObjectTable) *set) {
     if (!set) {
         set = ulib_alloc(set);
-        if (set) *set = uhset_init(CowlObjectTable);
+        if (set) *set = uhset(CowlObjectTable);
     }
     return (CowlIterator) { .ctx = set, .for_each = cowl_store_set };
 }
 
-CowlIterator cowl_iterator_count_init(ulib_uint *count) {
+CowlIterator cowl_iterator_count(ulib_uint *count) {
     if (!count) {
         count = ulib_alloc(count);
         if (count) *count = 0;